+++ src/pages/profile/profile.ts
@@ -58,14 +58,14 @@ export class ProfilePage {
     if (!this.uidUser){ // || this.userDetail.uid == this.uidUser
       this.profileYourself = true;
       this.currentUserDetail = this.chatManager.getLoggedUser();
-      this.displayImage(this.currentUserDetail.uid);
+      //this.displayImage(this.currentUserDetail.uid);
     }
     else {
       this.profileYourself = false;
       console.log('this.uidUser',this.uidUser);
       this.userDetail = new UserModel(this.uidUser, '', '', this.uidUser, '', '');
       this.userService.loadUserDetail(this.uidUser);
-      this.displayImage(this.uidUser);
+      //this.displayImage(this.uidUser);
     }
   }
   /**
@@ -87,28 +87,27 @@ export class ProfilePage {
   /**
    * carico url immagine profilo passando id utente
    */
-  displayImage(uidContact){
+  displayImage(uidContact: string){
     const that = this;
     console.log(" ********* displayImage::: ");
-    this.upSvc.display(uidContact, 'thumb')
-    .then(onResolve, onReject)
-    function onResolve(foundURL) {
-        console.log('foundURL', foundURL);
-        if(that.profileYourself){
-          that.currentUserDetail.imageurl = foundURL;
-        }else{
-          that.userDetail.imageurl = foundURL;
-        }
-    } 
-    function onReject(error){ 
-        console.log('error.code', error.code); 
-        const foundURL = 'assets/img/no_image.png';
-        if(that.profileYourself){
-          that.currentUserDetail.imageurl = foundURL;
-        }else{
-          that.userDetail.imageurl = foundURL;
-        }
-    }
+    // this.upSvc.display(uidContact, 'thumb')
+    // .then(onResolve, onReject)
+    // function onResolve(foundURL:string) {
+    //     console.log('foundURL', foundURL);
+    //     if(that.profileYourself){
+    //       that.currentUserDetail.imageurl = foundURL;
+    //     }else{
+    //       that.userDetail.imageurl = foundURL;
+    //     }
+    // } 
+    // function onReject(error:any){ 
+    //     console.log('error.code', error.code); 
+    //     if(that.profileYourself){
+    //       that.currentUserDetail.imageurl = '';
+    //     }else{
+    //       that.userDetail.imageurl = '';
+    //     }
+    // }
 
 
     // .then((url) => {