+++ src/utils/constants.ts
@@ -1,6 +1,6 @@
 //TIMES
 // BUILD
-export const CURR_VER_DEV = '1.075';
+export const CURR_VER_DEV = '1.087';
 export const CURR_VER_PROD = '0.934';
 
 //export const SERVICE_WORKER_DEV = '/chat/';